pyproject.toml
[tool.poetry]
name = "generic_agents"
version = "0.1.0"
description = "generic_agents using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = "^0.35.8" }
crewai-tools = "^0.4.6"
langchain-google-genai = "^1.0.3"
python-dotenv = "^1.0.1"
duckduckgo-search = "6.2.7"
langchain_groq = "0.1.3"

[tool.poetry.scripts]
generic_agents = "generic_agents.main:run"
train = "generic_agents.main:train"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
